Add completion options for send broadcast and start session
